api/requirements.txt
langchain==0.0.194
pinecone-client==2.2.1
kor==0.9.2
watchdog
fastapi==0.85.2
uvicorn==0.18.3
sse_starlette==1.3.3
python-multipart==0.0.6